requirements.txt
pandas==2.3.1
psycopg2-binary==2.9.10
openpyxl==3.1.5
